+++ Src/globals.h
@@ -4,12 +4,16 @@
 #include "stm32l476xx.h"
 
 
+typedef enum {false, true} bool;
+
 // Global Variables
 extern int core_MHz;
 extern int systick_time;
-extern int heartbeat_counter;
-extern int ag_counter;
 
+extern bool button0;
+extern bool button1;
+
+void nop(long long nop_loops);
 
 
 #endif // REALOP1_GLOBALS_H_
