--- Src/system_config/Interrupts/exti_config.h
@@ -1,17 +1,0 @@
-/*
- * exti_config.h
- *
- *  - May 14-16, 2023
- * 		Author       : Raphael, Darsh, Parteek
- *      Contributors : nithilsenthil, Huey, Raymond, Kevin
- *      Log          : wrote EXTI_init()
- */
-
-#ifndef REALOP1_SYSTEM_EXTI_CONFIG_H_
-#define REALOP1_SYSTEM_EXTI_CONFIG_H_
-
-#include "stm32l476xx.h"
-
-void EXTI_init();
-
-#endif /* REALOP1_SYSTEM_CONFIG_H_ */
