+++ Src/system_config/Timers/timers.h
@@ -1,0 +1,15 @@
+/*
+ * timers.h
+ */
+
+#ifndef REALOP1_TIMERS_H_
+#define REALOP1_TIMERS_H_
+
+#include "stm32l476xx.h"
+#include "../LED/led.h"
+#include "../../globals.h"
+
+void systick_init();
+
+
+#endif	// REALOP1_TIMERS_H_
